+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning].
 ## [Unreleased]
 
 - Add uninstall command to reverse install command
+- Allow output in JSON format
 
 ## [0.1.1] - 2023-11-25